app/requirements.txt
Django==4.1.7
asgiref==3.6.0
sqlparse==0.4.2
debugpy==1.5.1
mysqlclient==2.1.0
gunicorn==20.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
defusedxml==0.7.1
django-allauth==0.49.0
idna==3.3
oauthlib==3.2.0
pycparser==2.21
PyJWT==2.3.0
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
urllib3==1.26.8
djangorestframework==3.14.0
pyyaml==6.0
uritemplate==4.1.1
